Organized DiscordChessChallenger for clarity.
